\name{RandomSplitter-package}
\alias{RandomSplitter-package}
\alias{RandomSplitter}
\docType{package}
\title{\packageTitle{RandomSplitter}}
\description{\packageDescription{RandomSplitter}}
\section{Package Content}{\packageIndices{RandomSplitter}}
\author{\packageAuthor{RandomSplitter}}
\section{Maintainer}{\packageMaintainer{RandomSplitter}}
\keyword{package}
