% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomSplitter.R
\name{train_test_split}
\alias{train_test_split}
\title{Fast Train/Test Split Wrapper}
\usage{
train_test_split(df, train_size = 0.75, stratify_by = NULL)
}
\arguments{
\item{df}{A \code{data.frame} to split.}

\item{train_size}{Numeric. The proportion of data to be assigned to the
training set (between 0 and 1). Default is 0.75.}

\item{stratify_by}{Optional string. The name of the column to use for
stratified splitting.}
}
\value{
A named \code{list} containing two \code{data.frame} objects: \code{train} and \code{test}.
}
\description{
A convenience wrapper around \code{\link{split_df}} for the most common
machine learning use case: partitioning data into a training set and a testing set.
}
\examples{
# Simple 80/20 split
split_data <- train_test_split(mtcars, train_size = 0.8)
train_set <- split_data$train
test_set  <- split_data$test
}
\seealso{
\code{\link{split_df}}
}
